set default users in TeamEditor
