+++ src/components/domains/Team/TeamEditor/hooks.ts
@@ -17,6 +17,19 @@ export const useModal = () => {
 export const useField = (team?: Team) => {
   const [name, setName] = useState(team?.name ?? '');
   const [color, setColor] = useState(team?.color ?? '#000000');
-  const [users, setUsers] = useState<User[]>([]);
-  return { name, color, users, setName, setColor, setUsers };
+  return { name, color, setName, setColor };
+};
+
+export const useUserSelect = (users?: User[]) => {
+  const [selectedUsers, setSelectedUsers] = useState<User[]>(users || []);
+
+  const addUser = (user: User) => {
+    setSelectedUsers((prev) => prev.concat([user]));
+  };
+
+  const removeUser = (user: User) => {
+    setSelectedUsers((prev) => prev.filter((prevUser) => prevUser.id !== user.id));
+  };
+
+  return { selectedUsers, addUser, removeUser };
 };
